Cleanups

+++ src/device/environment.rs
@@ -137,8 +137,7 @@ impl Device {
                 marginal_cdf_floats[y] = f16::from_f32(marginal_cdf[y].cdf).to_bits();
             }
 
-            self.envmap_marg_cdf
-                .upload(rows as usize, 1, marginal_cdf_floats);
+            self.envmap_marg_cdf.upload(rows, 1, marginal_cdf_floats);
 
             // STEP 6: upload the conditional CDF to its own texture (one line
             // per CDF)
@@ -153,7 +152,7 @@ impl Device {
             }
 
             self.envmap_cond_cdf
-                .upload(cols as usize, rows as usize, conditional_cdf_floats);
+                .upload(cols, rows, conditional_cdf_floats);
         }
 
         Ok(())
